requirements.txt
quart
quart-rate-limiter
hypercorn
playwright
pydantic
Pillow
tenacity
ua-generator
python-dotenv
psutil
tldextract
validators
pytest-asyncio
pytest-cov
structlog
ua-generator